Got the local and server folders to mirror eachother as you click in and out of the directories. All buttons work.
There are some checks i didnt get a chance to work on:
How to handle an empty server but full local tree.
How to handle the changed ordering on the server side, so they stay changed.
check last modified dates to see if the directory or file needs to be modified.