Made index score query much more simple. Finished export to .csv file.
Added student view for the gradebook. Finished the Tab View and working on the Index View.

Added support for the students to enter grades when the professor hasn't done any yet. When the professor enters a grade for the student, they will no longer be able to enter a grade for that cell.