Got the Reply all working and put the buttons on top and bottom, both horizontal.
( old check in notes not sure if it got checked in before: Finally got the multi email working, I added an extra field to the mail model in the db.  It will be pretty much the same as before but you wont loose the Subject and Message fields on new emails either now and it is much more efficient, doesn't call server side everytime.  
I will add in buttons for the 3 sends and reply all features, just wanted to get this checked in. Sorry it took so long, i had to rework it like 3 times.)