Student Gradebook: made it so that if you drop more assignments than are in a category, it now calclulates the highest grade and displays it.
Student tab: Made it so that the average score now displays the average points instead of the average percent.

Teacher Gradebook: made it so when you select drop lowest where the value to drop is more than there are assignments it now displays the highest grade on submit instead of displaying NG until you modified a cell.

Gradebook: Verified with Chris that the gradebook should and does calculate the category totals based off of points instead of percentages.