Analytics has been updated with the new UI elements. Each type of view needs a controller that has an index and an options action that will return the correct partial view for it's respective displays.
NOTE: Calendar looks funny right now because of the fact that it has some structure errors in it with regards to bootstrap (rows inside columns, etc. ), and it has some absolute and static positioned elements which cause them to be drawn outside the bounds of the containing div in some cases. Calendar will need to be changed slightly to fit the new dynamic.