Fixed minor bugs: - Analytics tab was not being displayed as active when viewing the analytics page. - assignmentFilesUploader.js contained a bug where it was referencing an object that represented the old course select dropdown that no longer exists.
After creating every type of assignment with every option (Except ABET stuff) no other bugs were found. The ABET options were skipped because I don't really understand them and the message given on the ABET tab of the assignment wizard says to contact an administrator. I have no idea if this is  a bug or not.