AJ: Updated the way logging in is executed, now the FileCache, caches and AuthKey and saves that cookie in the user's browser when they have successfully logged in.  This was necessary to enable interaction with the OSBLE Plugin for VS.
NOTE: If students have issues logging in when this is pushed to the live server inform them all to clear their cookies, as we are using different cookies now for logging in.

Fixed a null reference exception whenclicking "View [student names]'s Discussion Posts" in a discussion assignment with a rubric.  Updated to use strings and Enum.TryParse, I was getting erratic behavior otherwise.