I added the ability for TAs and Instructors to be multi section users. Instructors are always in all sections, and TAs are in single sections, multiple sections, or all sections. I reworked the roster page to have a seperate script file rather than inline. I created a sweep method that enforces the new role regulations (students can't be in multiple sections and instructors are always in all sections). I reworked assignments index pages so that they only show students that are in your section/sections.
Database -- had to add a collumn to Course User called 'multiSection'. This is a comma separated string that shows all sections the user is in IF they are in section -1. section -2 is ALL sections, section -1 is multiple sections, and section greater than 0 are normal sections.

Ninja Fix -- Disabled Clone Assignment