Removed extraneous space from installer
The OSBLE Prerequisites Installer's name has a space preceding the actual name. This causes it to show up at the top of windows alphabetized add/remove programs menu.
The goal of the Commit is to remove that space, I was unable to fully build the project and test this change, so I am unsure if more are needed.